+++ corehq/apps/notifications/models.py
@@ -17,3 +17,24 @@ class Notification(models.Model):
 
     class Meta:
         ordering = ["-activated"]
+
+    @classmethod
+    def get_by_user(cls, user, limit=10):
+        notes = cls.objects.filter(is_active=True)[:limit]
+        read_notifications = cls.objects.filter(users_read=user)
+
+        def _fmt_note(note_idx):
+            index = note_idx[0]
+            note = note_idx[1]
+
+            note_dict = {
+                'id': note.id,
+                'url': note.url,
+                'date': '{dt:%B} {dt.day}'.format(dt=note.activated),
+                'content': note.content,
+                'type': note.type,
+                'isRead': (index > 4 or note in read_notifications),
+            }
+            return note_dict
+
+        return map(_fmt_note, enumerate(notes))
